requirements.txt
boto3==1.26.89
numpy==1.24.2
opencv-python-headless==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
tensorflow==2.11.0
